Hydro food delivery V3
This is the third version of my hydro food delivery app. It allows the user to add items to each restaurant, along with having slightly reworked/restructured layouts for the screen (Such as the header class not having a present height, and the description boxes having a smaller height). I also made it so that if text overflows (IE: Goes outside the box it's supposed to be stored in due to being too large), it cuts off and ends with "...". Which will help when handling restaurants/food items with longer names.