Hydro Food delivery V4
This is the 4th version of my hydro food delivery app. It has various changes to the structure to help make it look better, added a horizontal scroller for items with a delivery fee less and $2, items can now be deleted from menus and now when adding an item you can choose weather or not you wanna have them be chef special or vegetarian based on a checkbox.