Hydro Food Delivery V7
This is the 7th version of my Hydro Food delivery app. It adds the ability to log in, and the code is setup to only allow users to be able to add, edit or delete restaurants and/or items if the user is currently logged in as an admin user. Plus the homepage also displays the currently set delivery address in the top left corner when logged in.