Hydro Food Delivery app V8
This is the 8th version of my hydro food delivery app, it allows the user to login/sign out, along with only allowing the user to edit/delete items and/or restaurants if they're logged in ad are an admin user, along with being able to register as an admin user by entering the admin key. Admin users can also see all accounts and delete other accounts. There are also several small style changes, like the the stars for the review score are displayed to be 5 stars, being yellow or black depending on the score, buttons like the back button changing color when hovering over them, and item/restaurant cards moving when hovering the mouse over them.