Hydro Food Delivery App Final Version
This is the 9th and final version of my Hydro Food Delivery app. Outside of some small style changes, the only truly big change is that now, the name of the restaurant on the restaurant page no longer cuts off but is now smaller to avoid overlapping, and now when entering a name or adding/editing a restaurant, there's a 30-character limit.